refactor: move product stock render to render.js
